Review: have seprated files for SharedData and SharedQueue
