+++ src/shared_data.cc
@@ -1,0 +1,65 @@
+// Copyright 2020 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "src/shared_data.h"
+
+#include <deque>
+#include <map>
+#include <memory>
+#include <mutex>
+#include <unordered_map>
+#include <unordered_set>
+
+namespace proxy_wasm {
+
+SharedData global_shared_data;
+
+WasmResult SharedData::get(std::string_view vm_id, const std::string_view key,
+                           std::pair<std::string, uint32_t> *result) {
+  std::lock_guard<std::mutex> lock(mutex_);
+  auto map = data_.find(std::string(vm_id));
+  if (map == data_.end()) {
+    return WasmResult::NotFound;
+  }
+  auto it = map->second.find(std::string(key));
+  if (it != map->second.end()) {
+    *result = it->second;
+    return WasmResult::Ok;
+  }
+  return WasmResult::NotFound;
+}
+
+WasmResult SharedData::set(std::string_view vm_id, std::string_view key, std::string_view value,
+                           uint32_t cas) {
+  std::lock_guard<std::mutex> lock(mutex_);
+  std::unordered_map<std::string, std::pair<std::string, uint32_t>> *map;
+  auto map_it = data_.find(std::string(vm_id));
+  if (map_it == data_.end()) {
+    map = &data_[std::string(vm_id)];
+  } else {
+    map = &map_it->second;
+  }
+  auto it = map->find(std::string(key));
+  if (it != map->end()) {
+    if (cas && cas != it->second.second) {
+      return WasmResult::CasMismatch;
+    }
+    it->second = std::make_pair(std::string(value), nextCas());
+  } else {
+    map->emplace(key, std::make_pair(std::string(value), nextCas()));
+  }
+  return WasmResult::Ok;
+}
+
+} // namespace proxy_wasm
